remove cast

+++ evo/core/trajectory.py
@@ -447,7 +447,7 @@ def calc_angular_speed(p_1: np.ndarray, p_2: np.ndarray, t_1: float,
                                   str(t_2))
     angle_1 = lie.so3_log(p_1[:3, :3], degrees)
     angle_2 = lie.so3_log(p_2[:3, :3], degrees)
-    return float((angle_2 - angle_1) / (t_2 - t_1))
+    return (angle_2 - angle_1) / (t_2 - t_1)
 
 
 def xyz_quat_wxyz_to_se3_poses(
